Rename data model modules by convention
